style: add new config file for provider configuration
